+++ lib/nibble/fieldtypes.rb
@@ -3,7 +3,7 @@ module Nibble
     CORE = %w[
       Nibble::Fieldtypes::Text Nibble::Fieldtypes::Textarea Nibble::Fieldtypes::Integer Nibble::Fieldtypes::Toggle
       Nibble::Fieldtypes::Select Nibble::Fieldtypes::Radio Nibble::Fieldtypes::Checkboxes Nibble::Fieldtypes::Date
-      Nibble::Fieldtypes::Slug Nibble::Fieldtypes::RichText Nibble::Fieldtypes::Link Nibble::Fieldtypes::List Nibble::Fieldtypes::Seo
+      Nibble::Fieldtypes::Slug Nibble::Fieldtypes::RichText Nibble::Fieldtypes::Markdown Nibble::Fieldtypes::Link Nibble::Fieldtypes::List Nibble::Fieldtypes::Seo
       Nibble::Fieldtypes::Grid Nibble::Fieldtypes::Replicator Nibble::Fieldtypes::Assets Nibble::Fieldtypes::Entries Nibble::Fieldtypes::Terms
       Nibble::Fieldtypes::Secret Nibble::Fieldtypes::Files
     ].freeze